attack/patch_attack/PatchAttackTool/requirements.txt
addict==2.4.0
albumentations==1.1.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
google==3.0.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
imageio==2.15.0
joblib==1.1.0
kiwisolver==1.3.1
kornia==0.6.4
matplotlib==3.3.4
mmcv==1.4.8
networkx==2.5.1
numpy==1.18.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.3
qudida==0.0.4
requests==2.27.1
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.1.0
six==1.16.0
soupsieve==2.3.1
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.5.4
# torch==1.10.1+cu111
# torchaudio==0.10.1+rocm4.1
# torchvision==0.11.2+cu111
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
yacs==0.1.8
yapf==0.32.0
